adds tests for pure substance, taking co2 parameters hardcoded

+++ solution/src/main/java/com/gerg2008/app/model/Alpha_Ideal_oi.java
@@ -15,7 +15,7 @@ public class Alpha_Ideal_oi {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
     private String name;
-    private long k;
+    private int k;
     private Double n_oik;
     private Double teta_oik;
